refactor(core): nine config sections, nine headers, and a config change costs 55 % less
*** GPU GATES NOT RUN: the card is withheld. make dev and make dev-test (CPU)
*** are green and all static gates pass; make test-gpu and verify-fast are
*** OUTSTANDING. Local commit, not pushed.

MEASURED, and the reason this is not the split that was asked for.

  touch nothing (no-op)                              2.9 s
  edit logging.cpp, a small .cpp                     9.1 s
  edit kv_cache_manager.cpp, 1200 code LOC          14.2 s
  edit cuda_graph.cu, 1046 code LOC, a .cu          14.6 s
  add one field to dispatch_policy.h               137.1 s

The file-size gate polices .cu and .cpp code LOC, where the worst case is 14.6 s
against a 9.1 s floor - so splitting the largest translation unit four ways
buys about 4 s. Nothing gates header fan-out, and one config field costs 137.1 s.
CLAUDE.md's rationale for the size gate ("recompile blast radius, not line
count") is sound as reasoning and wrong about where the money is.

dispatch_policy.h is included by 23 TUs. 21 of them touch two or fewer of its
nine sections, and only 2 take the whole DispatchPolicy aggregate - and both of
those use exactly one section anyway. So the fan-out is not intrinsic.

This is F-10 one level down, and this header's own preamble records the
original: config.h was included by 22 files, 85 TUs transitively, and changed
130 times in six months, "the highest build cost in the repo". Lifting nine
sections into an aggregate fixed that and gave the aggregate the same property
for the same reason.

AFTER, adding one bool to speculative, three runs: 72.2 / 57.7 / 61.9 s, median
~62 s against 137.1 s. About 55 % less, from the pure move alone, with every
consumer still including the aggregate.

I cannot explain that. Every TU still reaches speculative.h transitively through
dispatch_policy.h, so on my model of ninja and ccache the set of rebuilt TUs
should be unchanged. The number is measured and reproduced three times; the
mechanism is NOT established and I am not claiming one.

PURE MOVE. Each section's contents are byte-identical; dispatch_policy.h
includes all nine and keeps the aggregate, so every existing
#include "core/dispatch_policy.h" compiles unchanged and no include is deleted
here. SwaSizingMode got its own header because it lives in namespace imp rather
than imp::cfg and kv_cache.h needs it.

NOT DONE, deliberately: no consumer switched to a narrow header. That is where
the rest of the win is and it is a separate change.

UNMEASURED: 8 of the 23 TUs appear to touch no section at all, but the query
that found that matched the words GEMM/MoE/GDN in comments, so whether they can
drop the include entirely is unknown. Not chased.
