+++ src/core/config/generation.h
@@ -1,0 +1,38 @@
+#pragma once
+
+// Generation configuration, one of the nine sections split out of
+// core/dispatch_policy.h on 2026-08-21.
+//
+// WHY. dispatch_policy.h aggregates all nine and is included by 23 translation
+// units, of which 21 touch two sections or fewer. Adding one field to it costs
+// 137.1 s of incremental rebuild, against 9.1 s for a small .cpp and 14.6 s for
+// the largest .cu the file-size gate polices. A TU that needs only this section
+// can include only this header and stop rebuilding when the others change.
+//
+// This is F-10 one level down, and dispatch_policy.h's own preamble records the
+// original: config.h was included by 22 files, 85 TUs transitively, and changed
+// 130 times in six months - "the highest build cost in the repo". Lifting nine
+// sections into an aggregate fixed that, and gave the aggregate the same
+// property for the same reason.
+//
+// Pure move: the contents below are byte-identical to their previous form, and
+// dispatch_policy.h includes every one of these, so no existing include breaks.
+
+#include <cstdint>
+#include <string>
+#include <vector>
+
+namespace imp {
+namespace cfg {
+
+struct Generation {
+    bool no_logit_softcap = false;
+    bool lm_dequant_fp16 = false;
+    bool force_bos = false;
+    // Disable banned-token list (debug).
+    bool no_ban = false;
+    // Disable RoPE inside the MTP draft head (diagnostic).
+    bool mtp_no_rope = false;
+};
+}  // namespace cfg
+}  // namespace imp
